+++ relay-base-schema/src/data_category.rs
@@ -71,14 +71,9 @@ pub enum DataCategory {
     SpanIndexed = 16,
     /// ProfileDuration
     ///
-    /// This data category is used to count the number of milliseconds we have per profile chunk.
-    ProfileDuration = 17,
-    /// ProfileDurationIndexed
-    ///
     /// This data category is used to count the number of milliseconds we have per indexed profile chunk.
     /// We will then bill per second.
-    ProfileDurationIndexed = 18,
-
+    ProfileDuration = 17,
     //
     // IMPORTANT: After adding a new entry to DataCategory, go to the `relay-cabi` subfolder and run
     // `make header` to regenerate the C-binding. This allows using the data category from Python.
@@ -112,7 +107,6 @@ impl DataCategory {
             "metric_bucket" => Self::MetricBucket,
             "span_indexed" => Self::SpanIndexed,
             "profile_duration" => Self::ProfileDuration,
-            "profile_duration_indexed" => Self::ProfileDurationIndexed,
             _ => Self::Unknown,
         }
     }
@@ -139,7 +133,6 @@ impl DataCategory {
             Self::MetricBucket => "metric_bucket",
             Self::SpanIndexed => "span_indexed",
             Self::ProfileDuration => "profile_duration",
-            Self::ProfileDurationIndexed => "profile_duration_indexed",
             Self::Unknown => "unknown",
         }
     }